Renamed add_entry to edit_entry and added edit functionality.
Edit entry now checks if an entry_id was passed in the URL to
see if it should edit an existing instance or create a new one.

Additionally, adding the implicit creator and group_id fields
are now handled in the view, when an empty Entry object is
created.

Fixed typo where title field wasn't getting set for invalid
forms.

Added entry_tags list to context to hold the tags currently
applied or in the process of being applied to the entry.
This list is used to decide which tags to show as active
in the template, and is drawn from either the entry's
stored tags (if editing) or the request data.