Moved all username validation into new user form class.
Instead of doing the username validation in the view
(e.g., checking to see if the username is already in use),
all username validation now happens in clean_username.

Also added a check for spaces in the username, since that
was previously allowed and caused weird bugs once it was
in the system.