Updated styling for entries in the list on the main page.
Added new style classes for the new "images only" and
"date grid" display modes.

Changed tags style class to "meta" since it now applies
not only to tags but also to the date created tagline.
Changed color to black because using the link color
was misleading to users (they're not clickable).