Reformatted and added options to the entry list header.
Updated display options from a toggling link that
switches between "list" and "images" to three display
option links which highlight whichever option is active.

Shortened the entry list heading and moved the edit link to the
end for clarity.

Added the date created to the bottom of expanded
entries underneath the tags.

Added two new display modes: images only and date grid.

Images only is a mode displays only the images in entries,
and thus only displays entries that have images. Each image
is a link to edit the entry, where you can see other info
like the title, tags, description, etc. It's designed for
image-based content, like a collection of icons.

Date grid (or calendar), is a mode that currently only
shows the title and the date created of each entry, in a
floating grid format. It's meant for a date-oriented overview
of a group, or for very short date-oriented entries (e.g.,
daily tarot readings).